requirements.txt
django==1.7.4
djangorestframework==3.0.5
django-cors-headers==1.0.0
djangorestframework-gis==0.8
